Modernize package setup
- Use Bun instead of npm and jest
- Use Biome instead of ESLint and Prettier
- Add TypeDoc
- Update TSConfig configuration
- Remove Authorizer
