--- .eslintrc.js
@@ -1,33 +1,0 @@
-/* eslint-disable unicorn/prefer-module */
-module.exports = {
-  root: true,
-  parser: "@typescript-eslint/parser",
-  plugins: [
-    "@typescript-eslint",
-    "unicorn",
-    "jest",
-    "import",
-    "prettier",
-    "promise",
-  ],
-  extends: [
-    "plugin:unicorn/recommended",
-    "plugin:@typescript-eslint/recommended",
-    "plugin:prettier/recommended",
-    "plugin:import/errors",
-    "plugin:import/warnings",
-    "plugin:import/typescript",
-    "plugin:promise/recommended",
-  ],
-  rules: {
-    "prefer-const": "off",
-    "@typescript-eslint/explicit-module-boundary-types": "off",
-    "@typescript-eslint/no-non-null-assertion": "off",
-    "no-unused-vars": "off",
-    "no-var": "off",
-    "unicorn/no-null": "off",
-    "unicorn/prefer-node-protocol": "off",
-    "unicorn/filename-case": "off",
-    "unicorn/prevent-abbreviations": "off",
-  },
-};
