--- .github/workflows/main.yml
@@ -1,82 +1,0 @@
-name: CI
-
-on:
-  push:
-    branches: main
-  pull_request: {}
-
-jobs:
-  build:
-    name: Build
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout repo
-        uses: actions/checkout@v2
-
-      - name: Use Node 18
-        uses: actions/setup-node@v1
-        with:
-          node-version: 18
-
-      - name: Install dependencies
-        uses: bahmutov/npm-install@v1
-
-      - name: Build
-        run: npm run build
-
-  typecheck:
-    name: Typechecker
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout repo
-        uses: actions/checkout@v2
-
-      - name: Use Node 18
-        uses: actions/setup-node@v1
-        with:
-          node-version: 18
-
-      - name: Install dependencies
-        uses: bahmutov/npm-install@v1
-
-      - name: Typecheck
-        run: npm run typecheck
-
-  test:
-    name: Unit and Integration Tests
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout repo
-        uses: actions/checkout@v2
-
-      - name: Use Node 18
-        uses: actions/setup-node@v1
-        with:
-          node-version: 18
-
-      - name: Install dependencies
-        uses: bahmutov/npm-install@v1
-
-      - name: Test
-        run: npm run test -- --ci --coverage --maxWorkers=2
-
-  lint:
-    name: Linter
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout repo
-        uses: actions/checkout@v2
-
-      - name: Use Node 18
-        uses: actions/setup-node@v1
-        with:
-          node-version: 18
-
-      - name: Install dependencies
-        uses: bahmutov/npm-install@v1
-
-      - name: Build
-        run: npm run build
-
-      - name: Lint
-        run: npm run lint
