+++ .github/workflows/publish.yml
@@ -5,16 +5,21 @@ on:
     types: [published]
 
 jobs:
-  build:
+  publish-npm:
+    name: "Publish to npm"
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/checkout@v1
-      - uses: actions/setup-node@v1
+      - uses: actions/checkout@v4
+      - uses: oven-sh/setup-bun@v2
+      - run: bun install --frozen-lockfile
+      - run: bun run build
+      - run: bun run exports
+
+      - uses: actions/setup-node@v4
         with:
-          node-version: 18
+          node-version: "lts/*"
           registry-url: https://registry.npmjs.org/
-      - run: npm install
-      - run: npm run build
+
       - run: npm publish --access public
         env:
           NODE_AUTH_TOKEN: ${{secrets.NPM_TOKEN}}
