+++ .gitignore
@@ -1,6 +1,3 @@
-/node_modules
 /build
 /coverage
-
-*.log
-.DS_Store
+/node_modules
