--- .prettierrc.js
@@ -1,4 +1,0 @@
-module.exports = {
-  singleQuote: false,
-  semi: true,
-}
