+++ biome.json
@@ -1,0 +1,44 @@
+{
+	"$schema": "https://biomejs.dev/schemas/1.7.1/schema.json",
+	"organizeImports": {
+		"enabled": true
+	},
+	"linter": {
+		"enabled": true,
+		"rules": {
+			"recommended": true,
+			"correctness": {
+				"useHookAtTopLevel": "error"
+			},
+			"performance": {
+				"noBarrelFile": "error",
+				"noReExportAll": "error"
+			},
+			"style": {
+				"noDefaultExport": "error",
+				"noNegationElse": "error",
+				"useConst": "off",
+				"useExportType": "off",
+				"useImportType": "off"
+			},
+			"suspicious": {
+				"noConsoleLog": "warn",
+				"noEmptyBlockStatements": "warn",
+				"noSkippedTests": "error"
+			}
+		}
+	},
+	"formatter": { "enabled": true },
+	"vcs": {
+		"enabled": true,
+		"clientKind": "git",
+		"defaultBranch": "main",
+		"useIgnoreFile": true
+	},
+	"overrides": [
+		{
+			"include": ["**/*.md"],
+			"formatter": { "indentStyle": "tab" }
+		}
+	]
+}
