+++ package.json
@@ -1,79 +1,62 @@
 {
-  "name": "remix-auth",
-  "version": "3.7.0",
-  "description": "Simple Authentication for Remix",
-  "main": "./build/index.js",
-  "types": "./build/index.d.ts",
-  "scripts": {
-    "build": "tsc --project tsconfig.json",
-    "typecheck": "tsc --project tsconfig.json --noEmit",
-    "lint": "eslint --ext .ts,.tsx src/",
-    "test": "jest --config=config/jest.config.ts --passWithNoTests",
-    "coverage": "npm run test -- --coverage"
-  },
-  "keywords": [
-    "remix",
-    "auth",
-    "authentication",
-    "local",
-    "auth0",
-    "oauth2",
-    "strategies"
-  ],
-  "author": {
-    "name": "Sergio Xalambrí",
-    "email": "hello@sergiodxa.com",
-    "url": "https://sergiodxa.com"
-  },
-  "repository": {
-    "url": "https://github.com/sergiodxa/remix-auth",
-    "type": "git"
-  },
-  "homepage": "https://github.com/sergiodxa/remix-auth#readme",
-  "license": "MIT",
-  "files": [
-    "build",
-    "package.json",
-    "README.md"
-  ],
-  "peerDependencies": {
-    "@remix-run/react": "^1.0.0 || ^2.0.0",
-    "@remix-run/server-runtime": "^1.0.0 || ^2.0.0"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.14.2",
-    "@babel/preset-env": "^7.14.1",
-    "@babel/preset-react": "^7.13.13",
-    "@babel/preset-typescript": "^7.13.0",
-    "@remix-run/node": "^2.0.1",
-    "@remix-run/react": "^2.0.1",
-    "@remix-run/serve": "^2.0.1",
-    "@remix-run/server-runtime": "^2.0.1",
-    "@types/jest": "^29.5.5",
-    "@types/prop-types": "^15.7.4",
-    "@types/react": "^18.2.20",
-    "@types/uuid": "^8.3.3",
-    "@typescript-eslint/eslint-plugin": "^6.7.3",
-    "@typescript-eslint/parser": "^6.7.3",
-    "babel-jest": "^26.6.3",
-    "eslint": "^7.26.0",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-cypress": "^2.11.3",
-    "eslint-plugin-import": "^2.22.1",
-    "eslint-plugin-jest": "^24.3.6",
-    "eslint-plugin-jest-dom": "^3.9.0",
-    "eslint-plugin-prettier": "^3.4.0",
-    "eslint-plugin-promise": "^5.1.0",
-    "eslint-plugin-testing-library": "^4.3.0",
-    "eslint-plugin-unicorn": "^32.0.1",
-    "jest": "^29.7.0",
-    "jest-fetch-mock": "^3.0.3",
-    "prettier": "^2.3.2",
-    "react": "^18.2.0",
-    "ts-node": "^9.1.1",
-    "typescript": "^5.1.6"
-  },
-  "dependencies": {
-    "uuid": "^8.3.2"
-  }
+	"name": "remix-auth",
+	"version": "3.7.0",
+	"description": "Simple Authentication for Remix",
+	"license": "MIT",
+	"funding": ["https://github.com/sponsors/sergiodxa"],
+	"author": {
+		"name": "Sergio Xalambrí",
+		"email": "hello+oss@sergiodxa.com",
+		"url": "https://sergiodxa.com"
+	},
+	"repository": {
+		"url": "https://github.com/sergiodxa/remix-auth",
+		"type": "git"
+	},
+	"homepage": "https://github.com/sergiodxa/remix-auth",
+	"bugs": {
+		"url": "https://github.com/sergiodxa/remix-auth/issues"
+	},
+	"scripts": {
+		"build": "tsc",
+		"typecheck": "tsc --noEmit",
+		"quality": "biome check .",
+		"quality:fix": "biome check . --write --unsafe",
+		"exports": "bun run ./scripts/exports.ts"
+	},
+	"sideEffects": false,
+	"type": "module",
+	"engines": {
+		"node": ">=20.0.0"
+	},
+	"files": ["build", "package.json", "README.md"],
+	"exports": {
+		".": "./build/index.js",
+		"./package.json": "./package.json"
+	},
+	"dependencies": {
+		"uuid": "^8.3.2"
+	},
+	"peerDependencies": {
+		"@remix-run/react": "^1.0.0 || ^2.0.0",
+		"@remix-run/server-runtime": "^1.0.0 || ^2.0.0"
+	},
+	"devDependencies": {
+		"@arethetypeswrong/cli": "^0.15.3",
+		"@babel/preset-react": "^7.13.13",
+		"@biomejs/biome": "^1.8.3",
+		"@remix-run/node": "^2.0.1",
+		"@remix-run/react": "^2.0.1",
+		"@remix-run/serve": "^2.0.1",
+		"@remix-run/server-runtime": "^2.0.1",
+		"@total-typescript/tsconfig": "^1.0.4",
+		"@types/bun": "^1.1.6",
+		"@types/react": "^18.2.20",
+		"@types/uuid": "^8.3.3",
+		"consola": "^3.2.3",
+		"react": "^18.2.0",
+		"typedoc": "^0.26.5",
+		"typedoc-plugin-mdn-links": "^3.2.6",
+		"typescript": "^5.5.4"
+	}
 }
