--- src/error.ts
@@ -1,5 +1,0 @@
-export class AuthorizationError extends Error {
-  constructor(message?: string, public cause?: Error) {
-    super(message);
-  }
-}
