+++ src/index.ts
@@ -1,4 +1,11 @@
-export * from "./authenticator";
-export * from "./authorizer";
-export * from "./error";
-export * from "./strategy";
+// biome-ignore lint/performance/noReExportAll: <explanation>
+// biome-ignore lint/performance/noBarrelFile: <explanation>
+export * from "./lib/authenticator.js";
+
+// biome-ignore lint/performance/noReExportAll: <explanation>
+// biome-ignore lint/performance/noBarrelFile: <explanation>
+export * from "./lib/error.js";
+
+// biome-ignore lint/performance/noReExportAll: <explanation>
+// biome-ignore lint/performance/noBarrelFile: <explanation>
+export * from "./lib/strategy.js";
