+++ src/lib/error.ts
@@ -1,0 +1,8 @@
+export class AuthorizationError extends Error {
+	constructor(
+		message?: string,
+		public override cause?: Error,
+	) {
+		super(message);
+	}
+}
