+++ tsconfig.json
@@ -1,16 +1,9 @@
 {
-  "compilerOptions": {
-    "lib": ["DOM", "DOM.Iterable", "ES2019"],
-    "esModuleInterop": true,
-    "moduleResolution": "Node",
-    "module": "CommonJS",
-    "target": "ES2019",
-    "strict": true,
-    "skipLibCheck": true,
-    "declaration": true,
-    "jsx": "react-jsx",
-    "outDir": "./build"
-  },
-  "exclude": ["node_modules"],
-  "include": ["src/**/*.ts", "src/**/*.tsx", "jest.d.ts"]
+	/* Change to `@total-typescript/tsconfig/tsc/dom/library` for DOM usage */
+	"extends": "@total-typescript/tsconfig/tsc/dom/library",
+	"include": ["src/**/*"],
+	"exclude": ["src/**/*.test.*"],
+	"compilerOptions": {
+		"outDir": "./build"
+	}
 }
