+++ typedoc.json
@@ -1,0 +1,10 @@
+{
+	"$schema": "https://typedoc.org/schema.json",
+	"includeVersion": true,
+	"entryPoints": ["./src/index.ts"],
+	"out": "pages",
+	"json": "pages/index.json",
+	"cleanOutputDir": true,
+	"plugin": ["typedoc-plugin-mdn-links"],
+	"categorizeByGroup": false
+}
